+++ tests/scroll.rs
@@ -1,6 +1,6 @@
 #[test]
 fn scroll_regions() {
-    let mut parser = vt100::Parser::new(24, 80, 0);
+    let mut parser = vt100::Parser::default();
     parser.process(b"\x1b[m\x1b[2J\x1b[H1\r\n2\r\n3\r\n4\r\n5\r\n6\r\n7\r\n8\r\n9\r\n10\r\n11\r\n12\r\n13\r\n14\r\n15\r\n16\r\n17\r\n18\r\n19\r\n20\r\n21\r\n22\r\n23\r\n24");
     assert_eq!(parser.screen().contents(), "1\n2\n3\n4\n5\n6\n7\n8\n9\n10\n11\n12\n13\n14\n15\n16\n17\n18\n19\n20\n21\n22\n23\n24");
 
@@ -40,7 +40,7 @@ fn scroll_regions() {
 
 #[test]
 fn origin_mode() {
-    let mut parser = vt100::Parser::new(24, 80, 0);
+    let mut parser = vt100::Parser::default();
 
     parser.process(b"\x1b[5;15r");
     assert_eq!(parser.screen().cursor_position(), (4, 0));
@@ -156,7 +156,7 @@ fn scrollback() {
 
 #[test]
 fn edge_of_screen() {
-    let mut parser = vt100::Parser::new(24, 80, 0);
+    let mut parser = vt100::Parser::default();
     let screen = parser.screen().clone();
 
     parser.process(b"\x1b[31m\x1b[24;75Hfooba\x08r\x08\x1b[1@a");